Еще оптимизировал и улучшил группы
